@@ -1,13 +1,12 @@
 package com.example.zjy.zjywidget.sample;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.zjy.zjywidget.R;
 import com.zjywidget.widget.fallingsurface.FallingSurfaceView;
-import com.zjywidget.widget.switchview.YSwitchView;
 
 /**
  * Created by 74215 on 2019/4/14.
