@@ -3,9 +3,9 @@ package com.example.zjy.zjywidget.sample;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.Nullable;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.zjy.zjywidget.R;
 import com.zjywidget.widget.waveload.YWaveLoadView;
